+++ poc/tests/errors_types/test_AllFplTypesRelatedErrorTests.py
@@ -25,6 +25,8 @@ class AllTypesRelatedErrorTests(UtilTestCase):
         ("test_FplTypeNotAllowed_ok_03.fpl", "SE0200"),
         ("test_FplTypeMismatch_ok_01.fpl", "SE0250"),
         ("test_FplTypeMismatch_ok_02.fpl", "SE0250"),
+        ("test_FplTypeMismatch_ok_03.fpl", "SE0250"),
+        ("test_FplTypeMismatch_ok_04.fpl", "SE0250"),
     ])
     def test_no_errors(self, use_case, diagnose_id):
         super().semantical_analysis_detects_no_fpl_error(self.folder + "/" + use_case, diagnose_id)